+++ src/algorithms/properties/mod.rs
@@ -31,7 +31,9 @@ pub(crate) mod reciprocity;
 pub(crate) mod triangles;
 
 pub use assortativity::{assortativity_degree, assortativity_degree_directed};
-pub use assortativity_weighted::assortativity_degree_weighted;
+pub use assortativity_weighted::{
+    assortativity_degree_directed_weighted, assortativity_degree_weighted,
+};
 pub use basic::{density, mean_distance};
 pub use betweenness::betweenness;
 pub use betweenness_weighted::betweenness_weighted;